:construction: Modify move function to work with PI
*

+++ inc/structs.h
@@ -6,7 +6,7 @@
 /*   By: owalsh <owalsh@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/27 09:16:02 by owalsh            #+#    #+#             */
-/*   Updated: 2022/11/17 16:57:27 by owalsh           ###   ########.fr       */
+/*   Updated: 2022/11/18 14:58:58 by owalsh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -95,7 +95,7 @@ typedef struct s_camera
 
 typedef struct s_player
 {
-	int			dir;
+	float		dir;
 	float		angle;
 	t_coord		pos;
 }				t_player;
